requirements.txt
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3.post1
chat-exporter==2.6.1
pillow==10.1.0